fix task 5 decimal ownership residuals

+++ src/main/java/com/toir/config/SampleDataSeeder.java
@@ -560,7 +560,7 @@ public class SampleDataSeeder implements CommandLineRunner {
         for (int i = 0; i < 8; i++) {
             ActualCost ac = new ActualCost();
             ac.setCostCategoryId(i % 2 == 0 ? materials.getId() : labor.getId());
-            ac.setAmount(50_000 + (long) (Math.random() * 200_000));
+            ac.setAmount(BigDecimal.valueOf(50_000 + (long) (Math.random() * 200_000)));
             ac.setCostDate(now.minus(i * 3L, ChronoUnit.DAYS));
             ac.setNotes("Фактические затраты по ремонту #" + (i + 1));
             ac.setStatus(i < 5 ? ActualCostStatus.APPROVED : ActualCostStatus.PENDING);
